Moving schema to stages
