+++ src/_nebari/stages/infrastructure/__init__.py
@@ -3,9 +3,19 @@ import inspect
 import pathlib
 import sys
 import tempfile
+import typing
 from typing import Any, Dict, List, Optional
 
+import pydantic
+
+from _nebari import constants
 from _nebari.stages.base import NebariTerraformStage
+from _nebari.provider.cloud import (
+    amazon_web_services,
+    azure_cloud,
+    digital_ocean,
+    google_cloud,
+)
 from _nebari.stages.tf_objects import (
     NebariAWSProvider,
     NebariGCPProvider,
@@ -20,10 +30,6 @@ def get_kubeconfig_filename():
     return str(pathlib.Path(tempfile.gettempdir()) / "NEBARI_KUBECONFIG")
 
 
-# TODO:
-# - create schema for node group for each provider
-
-
 class LocalInputVars(schema.Base):
     kubeconfig_filename: str = get_kubeconfig_filename()
     kube_context: Optional[str]
@@ -33,14 +39,20 @@ class ExistingInputVars(schema.Base):
     kube_context: str
 
 
-class BaseCloudProviderInputVars(schema.Base):
+class DigitalOceanNodeGroup(schema.Base):
+    instance: str
+    min_nodes: int
+    max_nodes: int
+
+
+class DigitalOceanInputVars(schema.Base):
     name: str
     environment: str
-    kubeconfig_filename: str = get_kubeconfig_filename()
-
-
-class DigitalOceanInputVars(BaseCloudProviderInputVars, schema.DigitalOceanProvider):
-    pass
+    region: str
+    tags: typing.List[str]
+    kubernetes_version: str
+    node_groups: typing.Dict[str, DigitalOceanNodeGroup]
+    kubeconfig_filename: str
 
 
 class GCPGuestAccelerators(schema.Base):
@@ -175,6 +187,315 @@ def kubernetes_provider_context(kubernetes_credentials: Dict[str, str]):
         yield
 
 
+class KeyValueDict(schema.Base):
+    key: str
+    value: str
+
+
+class DigitalOceanNodeGroup(schema.Base):
+    """Representation of a node group with Digital Ocean
+
+    - Kubernetes limits: https://docs.digitalocean.com/products/kubernetes/details/limits/
+    - Available instance types: https://slugs.do-api.dev/
+    """
+
+    instance: str
+    min_nodes: pydantic.conint(ge=1) = 1
+    max_nodes: pydantic.conint(ge=1) = 1
+
+
+class DigitalOceanProvider(schema.Base):
+    region: str = "nyc3"
+    kubernetes_version: typing.Optional[str]
+    # Digital Ocean image slugs are listed here https://slugs.do-api.dev/
+    node_groups: typing.Dict[str, DigitalOceanNodeGroup] = {
+        "general": DigitalOceanNodeGroup(
+            instance="g-8vcpu-32gb", min_nodes=1, max_nodes=1
+        ),
+        "user": DigitalOceanNodeGroup(
+            instance="g-4vcpu-16gb", min_nodes=1, max_nodes=5
+        ),
+        "worker": DigitalOceanNodeGroup(
+            instance="g-4vcpu-16gb", min_nodes=1, max_nodes=5
+        ),
+    }
+    tags: typing.Optional[typing.List[str]] = []
+
+    @pydantic.root_validator
+    def _validate_kubernetes_version(cls, values):
+        available_kubernetes_versions = digital_ocean.kubernetes_versions(
+            values["region"]
+        )
+        if (
+            values["kubernetes_version"] is not None
+            and values["kubernetes_version"] not in available_kubernetes_versions
+        ):
+            raise ValueError(
+                f"\nInvalid `kubernetes-version` provided: {values['kubernetes_version']}.\nPlease select from one of the following supported Kubernetes versions: {available_kubernetes_versions} or omit flag to use latest Kubernetes version available."
+            )
+        else:
+            values["kubernetes_version"] = available_kubernetes_versions[-1]
+        return values
+
+
+class GCPIPAllocationPolicy(schema.Base):
+    cluster_secondary_range_name: str
+    services_secondary_range_name: str
+    cluster_ipv4_cidr_block: str
+    services_ipv4_cidr_block: str
+
+
+class GCPCIDRBlock(schema.Base):
+    cidr_block: str
+    display_name: str
+
+
+class GCPMasterAuthorizedNetworksConfig(schema.Base):
+    cidr_blocks: typing.List[GCPCIDRBlock]
+
+
+class GCPPrivateClusterConfig(schema.Base):
+    enable_private_endpoint: bool
+    enable_private_nodes: bool
+    master_ipv4_cidr_block: str
+
+
+class GCPGuestAccelerator(schema.Base):
+    """
+    See general information regarding GPU support at:
+    # TODO: replace with nebari.dev new URL
+    https://docs.nebari.dev/en/stable/source/admin_guide/gpu.html?#add-gpu-node-group
+    """
+
+    name: str
+    count: pydantic.conint(ge=1) = 1
+
+
+class GCPNodeGroup(schema.Base):
+    instance: str
+    min_nodes: pydantic.conint(ge=0) = 0
+    max_nodes: pydantic.conint(ge=1) = 1
+    preemptible: bool = False
+    labels: typing.Dict[str, str] = {}
+    guest_accelerators: typing.List[GCPGuestAccelerator] = []
+
+
+class GoogleCloudPlatformProvider(schema.Base):
+    project: str = pydantic.Field(default_factory=lambda: os.environ["PROJECT_ID"])
+    region: str = "us-central1"
+    availability_zones: typing.Optional[typing.List[str]] = []
+    kubernetes_version: typing.Optional[str]
+    release_channel: str = constants.DEFAULT_GKE_RELEASE_CHANNEL
+    node_groups: typing.Dict[str, GCPNodeGroup] = {
+        "general": GCPNodeGroup(instance="n1-standard-8", min_nodes=1, max_nodes=1),
+        "user": GCPNodeGroup(instance="n1-standard-4", min_nodes=0, max_nodes=5),
+        "worker": GCPNodeGroup(instance="n1-standard-4", min_nodes=0, max_nodes=5),
+    }
+    tags: typing.Optional[typing.List[str]] = []
+    networking_mode: str = "ROUTE"
+    network: str = "default"
+    subnetwork: typing.Optional[typing.Union[str, None]] = None
+    ip_allocation_policy: typing.Optional[
+        typing.Union[GCPIPAllocationPolicy, None]
+    ] = None
+    master_authorized_networks_config: typing.Optional[
+        typing.Union[GCPCIDRBlock, None]
+    ] = None
+    private_cluster_config: typing.Optional[
+        typing.Union[GCPPrivateClusterConfig, None]
+    ] = None
+
+    @pydantic.root_validator
+    def _validate_kubernetes_version(cls, values):
+        available_kubernetes_versions = google_cloud.kubernetes_versions(
+            values["region"]
+        )
+        if (
+            values["kubernetes_version"] is not None
+            and values["kubernetes_version"] not in available_kubernetes_versions
+        ):
+            raise ValueError(
+                f"\nInvalid `kubernetes-version` provided: {values['kubernetes_version']}.\nPlease select from one of the following supported Kubernetes versions: {available_kubernetes_versions} or omit flag to use latest Kubernetes version available."
+            )
+        else:
+            values["kubernetes_version"] = available_kubernetes_versions[-1]
+        return values
+
+
+class AzureNodeGroup(schema.Base):
+    instance: str
+    min_nodes: int
+    max_nodes: int
+
+
+class AzureProvider(schema.Base):
+    region: str = "Central US"
+    kubernetes_version: typing.Optional[str]
+    node_groups: typing.Dict[str, AzureNodeGroup] = {
+        "general": AzureNodeGroup(instance="Standard_D8_v3", min_nodes=1, max_nodes=1),
+        "user": AzureNodeGroup(instance="Standard_D4_v3", min_nodes=0, max_nodes=5),
+        "worker": AzureNodeGroup(instance="Standard_D4_v3", min_nodes=0, max_nodes=5),
+    }
+    storage_account_postfix: str = pydantic.Field(
+        default_factory=lambda: random_secure_string(length=4)
+    )
+    vnet_subnet_id: typing.Optional[typing.Union[str, None]] = None
+    private_cluster_enabled: bool = False
+
+    @pydantic.validator("kubernetes_version")
+    def _validate_kubernetes_version(cls, value):
+        available_kubernetes_versions = azure_cloud.kubernetes_versions()
+        if value is None:
+            value = available_kubernetes_versions[-1]
+        elif value not in available_kubernetes_versions:
+            raise ValueError(
+                f"\nInvalid `kubernetes-version` provided: {value}.\nPlease select from one of the following supported Kubernetes versions: {available_kubernetes_versions} or omit flag to use latest Kubernetes version available."
+            )
+        return value
+
+
+class AWSNodeGroup(schema.Base):
+    instance: str
+    min_nodes: int = 0
+    max_nodes: int
+    gpu: bool = False
+    single_subnet: bool = False
+
+
+class AmazonWebServicesProvider(schema.Base):
+    region: str = pydantic.Field(
+        default_factory=lambda: os.environ.get("AWS_DEFAULT_REGION", "us-west-2")
+    )
+    availability_zones: typing.Optional[typing.List[str]]
+    kubernetes_version: typing.Optional[str]
+    node_groups: typing.Dict[str, AWSNodeGroup] = {
+        "general": AWSNodeGroup(instance="m5.2xlarge", min_nodes=1, max_nodes=1),
+        "user": AWSNodeGroup(
+            instance="m5.xlarge", min_nodes=1, max_nodes=5, single_subnet=False
+        ),
+        "worker": AWSNodeGroup(
+            instance="m5.xlarge", min_nodes=1, max_nodes=5, single_subnet=False
+        ),
+    }
+    existing_subnet_ids: typing.List[str] = None
+    existing_security_group_ids: str = None
+    vpc_cidr_block: str = "10.10.0.0/16"
+
+    @pydantic.root_validator
+    def _validate_kubernetes_version(cls, values):
+        available_kubernetes_versions = amazon_web_services.kubernetes_versions()
+        if values["kubernetes_version"] is None:
+            values["kubernetes_version"] = available_kubernetes_versions[-1]
+        elif values["kubernetes_version"] not in available_kubernetes_versions:
+            raise ValueError(
+                f"\nInvalid `kubernetes-version` provided: {values['kubernetes_version']}.\nPlease select from one of the following supported Kubernetes versions: {available_kubernetes_versions} or omit flag to use latest Kubernetes version available."
+            )
+        return values
+
+    @pydantic.root_validator
+    def _validate_availability_zones(cls, values):
+        if values["availability_zones"] is None:
+            zones = amazon_web_services.zones(values["region"])
+            values["availability_zones"] = list(sorted(zones))[:2]
+        return values
+
+
+class LocalProvider(schema.Base):
+    kube_context: typing.Optional[str]
+    node_selectors: typing.Dict[str, KeyValueDict] = {
+        "general": KeyValueDict(key="kubernetes.io/os", value="linux"),
+        "user": KeyValueDict(key="kubernetes.io/os", value="linux"),
+        "worker": KeyValueDict(key="kubernetes.io/os", value="linux"),
+    }
+
+
+class ExistingProvider(schema.Base):
+    kube_context: typing.Optional[str]
+    node_selectors: typing.Dict[str, KeyValueDict] = {
+        "general": KeyValueDict(key="kubernetes.io/os", value="linux"),
+        "user": KeyValueDict(key="kubernetes.io/os", value="linux"),
+        "worker": KeyValueDict(key="kubernetes.io/os", value="linux"),
+    }
+
+
+class InputSchema(schema.Base):
+    provider: schema.ProviderEnum = schema.ProviderEnum.local
+    local: typing.Optional[LocalProvider]
+    existing: typing.Optional[ExistingProvider]
+    google_cloud_platform: typing.Optional[GoogleCloudPlatformProvider]
+    amazon_web_services: typing.Optional[AmazonWebServicesProvider]
+    azure: typing.Optional[AzureProvider]
+    digital_ocean: typing.Optional[DigitalOceanProvider]
+
+    @pydantic.root_validator
+    def check_provider(cls, values):
+        if values["provider"] == schema.ProviderEnum.local and values.get("local") is None:
+            values["local"] = LocalProvider()
+        elif (
+            values["provider"] == schema.ProviderEnum.existing
+            and values.get("existing") is None
+        ):
+            values["existing"] = ExistingProvider()
+        elif (
+            values["provider"] == schema.ProviderEnum.gcp
+            and values.get("google_cloud_platform") is None
+        ):
+            values["google_cloud_platform"] = GoogleCloudPlatformProvider()
+        elif (
+            values["provider"] == schema.ProviderEnum.aws
+            and values.get("amazon_web_services") is None
+        ):
+            values["amazon_web_services"] = schema.AmazonWebServicesProvider()
+        elif values["provider"] == schema.ProviderEnum.azure and values.get("azure") is None:
+            values["azure"] = AzureProvider()
+        elif (
+            values["provider"] == schema.ProviderEnum.do
+            and values.get("digital_ocean") is None
+        ):
+            values["digital_ocean"] = DigitalOceanProvider()
+
+        if (
+            sum(
+                (_ in values and values[_] is not None)
+                for _ in {
+                    "local",
+                    "existing",
+                    "google_cloud_platform",
+                    "amazon_web_services",
+                    "azure",
+                    "digital_ocean",
+                }
+            )
+            != 1
+        ):
+            raise ValueError("multiple providers set or wrong provider fields set")
+        return values
+
+
+class NodeSelectorKeyValue(schema.Base):
+    key: str
+    value: str
+
+
+class KubernetesCredentials(schema.Base):
+    host: str
+    cluster_ca_certifiate: str
+    token: typing.Optional[str]
+    username: typing.Optional[str]
+    password: typing.Optional[str]
+    client_certificate: typing.Optional[str]
+    client_key: typing.Optional[str]
+    config_path: typing.Optional[str]
+    config_context: typing.Optional[str]
+
+
+class OutputSchema(schema.Base):
+    node_selectors: Dict[str, NodeSelectorKeyValue]
+    kubernetes_credentials: KubernetesCredentials
+    kubeconfig_filename: str
+    nfs_endpoint: typing.Optional[str]
+
+
 class KubernetesInfrastructureStage(NebariTerraformStage):
     """Generalized method to provision infrastructure.
 
@@ -193,6 +514,9 @@ class KubernetesInfrastructureStage(NebariTerraformStage):
 
     name = "02-infrastructure"
     priority = 20
+
+    input_schema = InputSchema
+    output_schema = OutputSchema
 
     @property
     def template_directory(self):
