+++ src/_nebari/stages/kubernetes_initialize/__init__.py
@@ -1,5 +1,8 @@
 import sys
+import typing
 from typing import Any, Dict, List, Union
+
+import pydantic
 
 from _nebari.stages.base import NebariTerraformStage
 from _nebari.stages.tf_objects import (
@@ -11,19 +14,57 @@ from nebari import schema
 from nebari.hookspecs import NebariStage, hookimpl
 
 
+class ExtContainerReg(schema.Base):
+    enabled: bool = False
+    access_key_id: typing.Optional[str]
+    secret_access_key: typing.Optional[str]
+    extcr_account: typing.Optional[str]
+    extcr_region: typing.Optional[str]
+
+    @pydantic.root_validator
+    def enabled_must_have_fields(cls, values):
+        if values["enabled"]:
+            for fldname in (
+                "access_key_id",
+                "secret_access_key",
+                "extcr_account",
+                "extcr_region",
+            ):
+                if (
+                    fldname not in values
+                    or values[fldname] is None
+                    or values[fldname].strip() == ""
+                ):
+                    raise ValueError(
+                        f"external_container_reg must contain a non-blank {fldname} when enabled is true"
+                    )
+        return values
+
+
 class InputVars(schema.Base):
     name: str
     environment: str
     cloud_provider: str
     aws_region: Union[str, None] = None
-    external_container_reg: Union[schema.ExtContainerReg, None] = None
+    external_container_reg: Union[ExtContainerReg, None] = None
     gpu_enabled: bool = False
     gpu_node_group_names: List[str] = []
+
+
+class InputSchema(schema.Base):
+    external_container_reg: ExtContainerReg = ExtContainerReg()
+
+
+class OutputSchema(schema.Base):
+    pass
 
 
 class KubernetesInitializeStage(NebariTerraformStage):
     name = "03-kubernetes-initialize"
     priority = 30
+
+    input_schema = InputSchema
+    output_schema = OutputSchema
 
     def tf_objects(self) -> List[Dict]:
         return [
