+++ src/_nebari/subcommands/validate.py
@@ -30,8 +30,5 @@ def nebari_subcommand(cli: typer.Typer):
             pass
         else:
             from nebari.plugins import nebari_plugin_manager
-
-            config_schema = nebari_plugin_manager.config_schema
-
-            schema.read_configuration(config_filename, config_schema=config_schema)
+            schema.read_configuration(config_filename, config_schema=nebari_plugin_manager.config_schema)
             print("[bold purple]Successfully validated configuration.[/bold purple]")
