+++ src/nebari/hookspecs.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List
 
 import typer
 from pluggy import HookimplMarker, HookspecMarker
-from pydantic import BaseModel
+import pydantic
 
 from nebari import schema
 
@@ -15,7 +15,9 @@ hookimpl = HookimplMarker("nebari")
 class NebariStage:
     name: str = None
     priority: int = None
-    stage_schema: BaseModel = None
+
+    input_schema: pydantic.BaseModel = None
+    output_schema: pydantic.BaseModel = None
 
     def __init__(self, output_directory: pathlib.Path, config: schema.Main):
         self.output_directory = output_directory
